Split it into chunks of 1000 movies each
When I ran the initial file, my PC hung for quite a while, and could not
see what the program was doing. So added a bit of text output to track
the progress and writing files for every 1000 movies, that way we will
have 18 files. Will post the files once all are ready.